changed directory structure; included seqan

+++ src/clone.cpp
@@ -96,7 +96,7 @@ cerr << "\nGenerating genome for " << *this << ", " << this->m_vec_mutations.siz
 void Clone::applyMutations(const vector<Mutation> &my_mutations,
                           SubstitutionModel model,
                           Genome &my_genome,
-                          boost::function<float()>& rng)
+                          boost::function<double()>& rng)
 {
   // sort mutations by position
   std::vector<Mutation> mut_sorted = Mutation::sortByPosition(my_mutations);
